requirements.txt
groq==0.13.*
flask==3.0.*
python-dotenv==1.0.*
python-docx==1.1.*
rank-bm25==0.2.*
orjson==3.10.*
regex==2024.11.*
uvicorn==0.30.*
chromadb==0.5.*
sentence-transformers==3.1.*
